# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Frontier Intel API - EVE Frontier Galaxy Intelligence Dashboard
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: kills
    description: Kill event data from on-chain telemetry indexer
  - name: assemblies
    description: Smart Assembly structure data from on-chain telemetry indexer
  - name: jumps
    description: Gate jump traffic data from on-chain telemetry indexer
  - name: intel
    description: Player-submitted intel reports
  - name: telemetry
    description: On-chain telemetry indexer operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /kills:
    get:
      operationId: getKills
      tags: [kills]
      summary: Get recent kill events
      description: Fetches indexed kill events from local on-chain telemetry store
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: solar_system_id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of kill events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KillsResponse"

  /assemblies:
    get:
      operationId: getAssemblies
      tags: [assemblies]
      summary: Get smart assemblies
      description: Fetches indexed player-built structures from local on-chain telemetry store
      parameters:
        - name: solar_system_id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of smart assemblies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssembliesResponse"

  /jumps:
    get:
      operationId: getJumps
      tags: [jumps]
      summary: Get gate jump traffic
      description: Fetches indexed gate jump events from local on-chain telemetry store
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 200
      responses:
        "200":
          description: Gate jump traffic data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JumpsResponse"

  /intel:
    get:
      operationId: getIntelReports
      tags: [intel]
      summary: Get all intel reports
      description: Returns all player-submitted intel reports
      parameters:
        - name: solar_system_id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of intel reports
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntelReportsResponse"
    post:
      operationId: createIntelReport
      tags: [intel]
      summary: Submit an intel report
      description: Submit a wallet-signed intel report for a solar system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIntelReportRequest"
      responses:
        "201":
          description: Intel report created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntelReport"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /intel/{id}:
    delete:
      operationId: deleteIntelReport
      tags: [intel]
      summary: Delete an intel report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /systems:
    get:
      operationId: getSolarSystems
      tags: [kills]
      summary: Get solar system threat summary
      description: Returns aggregated threat scores for all solar systems
      responses:
        "200":
          description: Solar system threat map data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SolarSystemsResponse"

  /telemetry/sync:
    post:
      operationId: syncTelemetry
      tags: [telemetry]
      summary: Sync telemetry indexer from chain
      description: Queries Sui events by configured package/module sources and stores normalized telemetry rows.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                limit:
                  type: integer
                  minimum: 1
                  maximum: 250
                  default: 100
      responses:
        "200":
          description: Telemetry sync result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelemetrySyncResponse"

  /telemetry/status:
    get:
      operationId: getTelemetryStatus
      tags: [telemetry]
      summary: Get telemetry indexer status
      description: Returns cursor and source configuration status for the custom indexer.
      responses:
        "200":
          description: Telemetry indexer status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelemetryIndexerStatusResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    KillEvent:
      type: object
      properties:
        killmail_id:
          type: string
        killer_character_id:
          type: string
        victim_character_id:
          type: string
        solar_system_id:
          type: string
        loss_type:
          type: string
          enum: [SHIP, STRUCTURE]
        kill_timestamp:
          type: string
      required:
        - killmail_id
        - killer_character_id
        - victim_character_id
        - solar_system_id
        - loss_type
        - kill_timestamp

    KillsResponse:
      type: object
      properties:
        kills:
          type: array
          items:
            $ref: "#/components/schemas/KillEvent"
        total:
          type: integer
        gateway_status:
          type: string
      required:
        - kills
        - total
        - gateway_status

    SmartAssembly:
      type: object
      properties:
        assembly_id:
          type: string
        solar_system_id:
          type: string
        owner_character_id:
          type: string
        assembly_type:
          type: string
          enum: [GATE, TURRET, STORAGE_UNIT, NETWORK_NODE, UNKNOWN]
        is_online:
          type: boolean
        name:
          type: string
      required:
        - assembly_id
        - solar_system_id
        - owner_character_id
        - assembly_type
        - is_online

    AssembliesResponse:
      type: object
      properties:
        assemblies:
          type: array
          items:
            $ref: "#/components/schemas/SmartAssembly"
        total:
          type: integer
        gateway_status:
          type: string
      required:
        - assemblies
        - total
        - gateway_status

    JumpEvent:
      type: object
      properties:
        character_id:
          type: string
        from_solar_system_id:
          type: string
        to_solar_system_id:
          type: string
        jump_timestamp:
          type: string
        gate_id:
          type: string
      required:
        - character_id
        - from_solar_system_id
        - to_solar_system_id
        - jump_timestamp

    JumpsResponse:
      type: object
      properties:
        jumps:
          type: array
          items:
            $ref: "#/components/schemas/JumpEvent"
        total:
          type: integer
        gateway_status:
          type: string
      required:
        - jumps
        - total
        - gateway_status

    IntelReport:
      type: object
      properties:
        id:
          type: integer
        solar_system_id:
          type: string
        message:
          type: string
        wallet_address:
          type: string
        report_type:
          type: string
          enum: [FLEET_SPOTTED, AMBUSH, SAFE, TRADE_ROUTE, OTHER]
        created_at:
          type: string
        expires_at:
          type: string
      required:
        - id
        - solar_system_id
        - message
        - wallet_address
        - report_type
        - created_at

    CreateIntelReportRequest:
      type: object
      properties:
        solar_system_id:
          type: string
        message:
          type: string
        wallet_address:
          type: string
        report_type:
          type: string
          enum: [FLEET_SPOTTED, AMBUSH, SAFE, TRADE_ROUTE, OTHER]
        signature:
          type: string
      required:
        - solar_system_id
        - message
        - wallet_address
        - report_type

    IntelReportsResponse:
      type: object
      properties:
        reports:
          type: array
          items:
            $ref: "#/components/schemas/IntelReport"
        total:
          type: integer
      required:
        - reports
        - total

    SolarSystemThreat:
      type: object
      properties:
        solar_system_id:
          type: string
        threat_level:
          type: string
          enum: [HIGH, MEDIUM, LOW, UNKNOWN]
        kill_count_1h:
          type: integer
        kill_count_24h:
          type: integer
        jump_count_1h:
          type: integer
        assembly_count:
          type: integer
        intel_count:
          type: integer
        last_activity:
          type: string
      required:
        - solar_system_id
        - threat_level
        - kill_count_1h
        - kill_count_24h
        - jump_count_1h
        - assembly_count
        - intel_count

    SolarSystemsResponse:
      type: object
      properties:
        systems:
          type: array
          items:
            $ref: "#/components/schemas/SolarSystemThreat"
        total:
          type: integer
        last_updated:
          type: string
      required:
        - systems
        - total
        - last_updated

    TelemetrySyncSourceSummary:
      type: object
      properties:
        stream:
          type: string
          enum: [kills, jumps, assemblies]
        package_id:
          type: string
        module:
          type: string
        synced_count:
          type: integer
        skipped_count:
          type: integer
        next_cursor:
          type: object
          nullable: true
          additionalProperties: true
        has_next_page:
          type: boolean
      required:
        - stream
        - package_id
        - module
        - synced_count
        - skipped_count
        - has_next_page

    TelemetrySyncResponse:
      type: object
      properties:
        rpc_url:
          type: string
        package_ids:
          type: array
          items:
            type: string
        summaries:
          type: array
          items:
            $ref: "#/components/schemas/TelemetrySyncSourceSummary"
        total_synced:
          type: integer
        total_skipped:
          type: integer
        synced_at:
          type: string
      required:
        - rpc_url
        - package_ids
        - summaries
        - total_synced
        - total_skipped
        - synced_at

    TelemetryIndexerCursorStatus:
      type: object
      properties:
        stream_key:
          type: string
        updated_at:
          type: string
        has_cursor:
          type: boolean
      required:
        - stream_key
        - updated_at
        - has_cursor

    TelemetryIndexerStatusResponse:
      type: object
      properties:
        rpc_url:
          type: string
        package_ids:
          type: array
          items:
            type: string
        cursor_count:
          type: integer
        cursors:
          type: array
          items:
            $ref: "#/components/schemas/TelemetryIndexerCursorStatus"
        checked_at:
          type: string
      required:
        - rpc_url
        - package_ids
        - cursor_count
        - cursors
        - checked_at

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
      required:
        - error

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success